Tiny URL Project Complete! (#6)
* new branch for URL shortener project
* added app.js, git ignore, and package.json

* updated package.json for URL shortener to include start script

* foundation laid:
* models: created a model for each URL created
* tools: cleaner [cleans the database for docs matching criteria], dbQuery [all interactions with db], shortener [shortens links and creates a new doc to be passed to db]
* controller: basics put in place
* app: basics put in place

* added a new schema for storageArray
* now stores all URLs in an array. easier to iterate and add to

* built up tools and database formatting:
* tools: added/updated shortener module: newShortURL, createCollection, addURL, originalURL functions
* DB_Format: added a URL array schema which accepts the shortenedURL schema for array elements
* tested using a local TEST_Shortener. all functions work as expected. added this file to gitignore to keep it local

* new tool: cleaner
* tools/cleaner.js : iterates through the array and checks each document against the criteria. if the document fails it is deleted from the array
to prevent buildup of old / unused shortenedURLs in the database

* Removed package-lock

* Removed package-lock

* Removed package-lock

* cleaned up code

* redirect are live                                                                                           [1:51:33]
* added an exporter file to tools which exports each tool individually for use elsewhere in the app
* updated the controller to include a redirect based on the ID parameter of the request. local testing works correctly. will eventually function
by testing the ID parameter of the request against a matching route from a document in the collection, if a match is found the original URL is
passed into res.redirect to send the user to their expected (long form) url

* made a mess:
made a mess in dbQuery, controller, commands. cant seem to be able to find the existing collection in the database. will look again tomorrow

* cleaned up code. error persists

* fixed the issue:

it was missing a [0] on the returned document from using the find mongoose method. working as intended

need to design a front end that posts to the server to shorten the url then pass it back.

* moved mongoose connect to controller to fix process.env missing bug

* prepared code for tomorrow's work

* changes to URL_Array Schema and cleaner function
* URL_Array: added a new field [updated] which tracks when the URL Array was last updated by the cleaner
* cleaner: checks if it has been 24 or more hours since the cleaner was run. if so run the cleaner

* began work on EJS view
view will allow user to paste their link --> submit --> and have the view updated with the shortened URL

* links and rendering:
posting a url through the (soon to be created) form will call a chain of events (setURL --> addURL --> return the shortened url and render the link page with it

* added clipboard package:
still configuring it, it will allow the user to copy the link to their clipboard.
either by button or on page load (could be annoying if unexpected)

* added clipboard button
updated the file structure to hold static files in /public and serve them in link.ejs
link.ejs renders being passed the shortened URL and places it in an input box. a clipboard icon next to the input box
can be clicked which automatically copies the link to the user's clipboard.
thanks to clipboard.js fantastic tool

* Completed Project!
* index.ejs will now load and be automatically re-rendered with the shortened URL when the user submits (posts) their long URL
* fixed route creation
* cleaned up code
ready for deployment